+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 0.1
 
 - start-rate limit support
+- renamed `limit` to `concurrency` (leaving old field, marked deprecated)